Magnus

+++ Social_medie_projekt/WebApi/Controllers/PostController.cs
@@ -12,7 +12,7 @@
             _postService = postService;
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpGet]
         public async Task<IActionResult> GetAllPostsAsync()
         {
@@ -20,7 +20,7 @@
             {
                 List<PostResponse> posts = await _postService.GetAllPostsAsync();
 
-                if (posts.Count() == 0)
+                if (posts.Count == 0)
                 {
                     return NoContent();
 
@@ -33,7 +33,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpGet]
         [Route("{postId}")]
         public async Task<IActionResult> GetPostByPostIdAsync([FromRoute] int postId)
@@ -54,7 +54,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpGet]
         [Route("user/{userId}")]
         public async Task<IActionResult> GetPostByUserIdAsync([FromRoute] int userId)
@@ -75,12 +75,19 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpPost]
         public async Task<IActionResult> CreatePostAsync([FromBody] PostRequest newPost)
         {
             try
             {
+                LoginResponse? currentUser = (LoginResponse?)HttpContext.Items["User"];
+
+                if (currentUser != null && newPost.UserId != currentUser.User.UserId)
+                {
+                    return Unauthorized(new { message = "Unauthorized" });
+                }
+
                 var postResponse = await _postService.CreatePostAsync(newPost);
 
                 return Ok(postResponse);
@@ -114,13 +121,20 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpDelete]
         [Route("{postId}")]
         public async Task<IActionResult> DeletePostAsync([FromRoute] int postId)
         {
             try
             {
+                LoginResponse? currentUser = (LoginResponse?)HttpContext.Items["User"];
+
+                if (currentUser != null && !currentUser.User.Posts.Exists(x => x.PostId == postId) && currentUser.Type != Role.admin)
+                {
+                    return Unauthorized(new { message = "Unauthorized" });
+                }
+
                 var postResponse = await _postService.DeletePostAsync(postId);
 
                 if (postResponse == null)
@@ -136,7 +150,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpPost]
         [Route("Like")]
         public async Task<IActionResult> CreateLikeAsync([FromBody] LikedRequest newLike)
@@ -153,7 +167,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpDelete]
         [Route("Like")]
         public async Task<IActionResult> DeleteLikeAsync([FromBody] LikedRequest deleteLike)
@@ -175,7 +189,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpGet]
         [Route("Tag")]
         public async Task<IActionResult> GetAllTagsAsync()
@@ -184,10 +198,9 @@
             {
                 List<TagResponse> tags = await _postService.GetAllTagsAsync();
 
-                if (tags.Count() == 0)
+                if (tags.Count == 0)
                 {
                     return NoContent();
-
                 }
                 return Ok(tags);
             }
@@ -197,7 +210,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpGet]
         [Route("Tag/{tagId}")]
         public async Task<IActionResult> GetTagByIdAsync([FromRoute] int tagId)
@@ -218,7 +231,7 @@
             }
         }
 
-        [Authorize(Role.user)]
+        [Authorize(Role.user, Role.admin)]
         [HttpPost]
         [Route("Tag")]
         public async Task<IActionResult> CreateTagAsync([FromBody] TagRequest newTag)
